refactor(monitor_utils): refactor monitor_utils as a class
Construct class build_runner instead of use a long function. Now
process_utils only contains stateless utils.

+++ src/mymonitor/data_models.py
@@ -135,7 +135,7 @@ class RunContext:
 
     # --- Run-specific parameters ---
     parallelism_level: int
-    monitoring_interval: int
+    monitoring_interval: float
     collector_type: str
     current_timestamp_str: str
